fix(json): fail the command a value cannot be JSON-encoded instead of storing an empty one
`JsonRef::serialize` swallowed the `serde_json` error and answered
`serialize_unit()`, so the command was still built and still sent with a
unit argument where the value belonged. `client.set(key, JsonRef(&value))`
returned `Ok(())` having written an absent value under the caller's key —
the mirror image of the fast-path panic, and the harder half of the family
to notice, because nothing fails.

The error now propagates and reaches the caller as
`ClientError::SerdeSerialize` before anything is sent, through the deferred
error slot the generic builder already had.

Three more results of the same pass:

- `Json<T>` implements `Serialize`, and `JsonRef` is removed. `Json(&value)`
  borrows exactly as `JsonRef(&value)` did, since `&T` is itself
  `Serialize`, which made the second type redundant rather than
  complementary.
- The visitor no longer answers `invalid type: Option` on a nil reply: it
  names `Option<Json<T>>`, and the integer, double and boolean arms name the
  reply shape they got. The transient-input arms are diagnosed rather than
  supported, since accepting them would force `T: DeserializeOwned`.
- `JsonArrIndexOptions::start` and `stop` are `isize`. A negative `start` is
  an offset from the end, verified against the server, and `u32` could not
  express it; both doc-comments described `stop` backwards.

Covered by twelve unit tests needing no server and by live tests closing the
option surface no test reached: `XX`, `From<SetCondition>`, the server's
rejection of the value-comparison conditions, `FP32`/`FP64`, a multi-path
`JSON.GET` and `NEWLINE` alone.

+++ src/commands/json_commands.rs
@@ -707,27 +707,29 @@ pub enum JsonFpType {
 #[serde(rename_all(serialize = "UPPERCASE"))]
 pub struct JsonArrIndexOptions {
     #[serde(rename = "", skip_serializing_if = "Option::is_none")]
-    start: Option<u32>,
+    start: Option<isize>,
     #[serde(rename = "", skip_serializing_if = "Option::is_none")]
-    stop: Option<i32>,
+    stop: Option<isize>,
 }
 
 impl JsonArrIndexOptions {
-    /// Inclusive start value to specify in a slice of the array to search.
+    /// Inclusive index of the first element of the slice to search.
     ///
-    /// Default is 0.
+    /// Default is 0, the first element.
+    /// Negative values are interpreted as starting from the end.
     #[must_use]
-    pub fn start(mut self, start: u32) -> Self {
+    pub fn start(mut self, start: isize) -> Self {
         self.start = Some(start);
         self
     }
 
-    /// Exclusive stop value to specify in a slice of the array to search, including the last element.
+    /// Exclusive index of the end of the slice to search.
     ///
-    /// Default is 0.
-    /// Negative values are interpreted as starting from the end.
+    /// Default is 0, the one value that means "to the end of the array".
+    /// Negative values are interpreted as starting from the end and stay
+    /// exclusive, so -1 drops the last element.
     #[must_use]
-    pub fn stop(mut self, stop: i32) -> Self {
+    pub fn stop(mut self, stop: isize) -> Self {
         self.stop = Some(stop);
         self
     }
